catalog/tools/smart_search_tool/requirements.txt
# Dependências do Google ADK
google-adk
